feat: Add zip_resources to gopackager_compile data source and update aws_lambda_function resource
This commit adds the `zip_resources` attribute to the `gopackager_compile` data source in the `data-source.tf` file. It also updates the `aws_lambda_function` resource in the same file to include the `source_code_hash` attribute, which is set to the base64-encoded SHA256 hash of the output_git_hash.